requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
lightgbm
streamlit
streamlit-folium
streamlit-autorefresh
folium
mlflow
sqlalchemy
psycopg2-binary
schedule
joblib
pillow
torch
torchvision